move multi-index logic to stores + clean-up
